[#12] refactor: 통합되는 ResponseDto 추가
