--- src/main/java/com/jeju/nanaland/domain/member/dto/response/LoginResponse.java
@@ -1,20 +1,0 @@
-package com.jeju.nanaland.domain.member.dto.response;
-
-import com.jeju.nanaland.domain.common.dto.response.ImageFileResponse;
-import com.jeju.nanaland.domain.common.dto.response.LanguageResponse;
-import com.jeju.nanaland.global.jwt.dto.JwtResponse;
-import lombok.Builder;
-import lombok.Getter;
-
-@Builder
-@Getter
-public class LoginResponse {
-
-  private JwtResponse jwtResponse;
-  private LanguageResponse languageResponse;
-  private ImageFileResponse imageFileResponse;
-  private Long memberId;
-  private String email;
-  private String nickname;
-  private String description;
-}
